stcompact: Exposing generation to communication (CompactReport)
