+++ components/stcompact/src/compact_node/convert.rs
@@ -276,6 +276,7 @@ impl From<persist::OpenPayment> for OpenPayment {
             dest_public_key: from.dest_public_key,
             dest_payment: from.dest_payment,
             description: from.description,
+            generation: from.generation,
             status: from.status.into(),
         }
     }
@@ -287,6 +288,7 @@ impl From<persist::OpenInvoice> for OpenInvoice {
             currency: from.currency,
             total_dest_payment: from.total_dest_payment,
             description: from.description,
+            generation: from.generation,
         }
     }
 }
